Adding my bot to a remote repository so I can try and host it online
I have been working on this bot for quite some weeks now as a backup to my friend's bot MechaSenku. It's main purpose right now is to use it in a way to test new features before implementing them into MechaSenku.

I was testing using this bot by running it locally in my machine, but as more time passed I realized that by now I should have automated the process a but and that is exactly what I'm trying to do right now. So instead of running it locally, I'll do what the cool kids do these days and host it online to make my life easier and learn new things as I go on.